Keep server prefs update in one place.
Makes sure all necessary keys are written and data stays consistent.
Previously, primary server wasn't updated when adding the first server,
and server ID set as well as active and primary servers weren't written
at well when adding a discovered server.
